remove old williamson5 script
